fix: more progress
