+++ src/routes/_static/wordFilters.js
@@ -1,0 +1,13 @@
+export const WORD_FILTER_CONTEXT_HOME = 'home'
+export const WORD_FILTER_CONTEXT_NOTIFICATIONS = 'notifications'
+export const WORD_FILTER_CONTEXT_PUBLIC = 'public'
+export const WORD_FILTER_CONTEXT_THREAD = 'thread'
+export const WORD_FILTER_CONTEXT_ACCOUNT = 'account'
+
+export const WORD_FILTER_CONTEXTS = [
+  WORD_FILTER_CONTEXT_HOME,
+  WORD_FILTER_CONTEXT_NOTIFICATIONS,
+  WORD_FILTER_CONTEXT_PUBLIC,
+  WORD_FILTER_CONTEXT_THREAD,
+  WORD_FILTER_CONTEXT_ACCOUNT
+]
