+++ src/routes/_store/computations/wordFilterComputations.js
@@ -1,6 +1,5 @@
 import { createRegexFromFilters } from '../../_utils/createRegexFromFilters'
-
-const FILTER_TYPES = ['home', 'notifications', 'public', 'thread', 'account']
+import { WORD_FILTER_CONTEXTS } from '../../_static/wordFilters'
 
 export function wordFilterComputations (store) {
   store.compute(
@@ -14,7 +13,7 @@ export function wordFilterComputations (store) {
       return null
     }
     const res = {}
-    for (const filterType of FILTER_TYPES) {
+    for (const filterType of WORD_FILTER_CONTEXTS) {
       res[filterType] = createRegexFromFilters(currentFilters.filter(_ => _.context.includes(filterType)))
     }
     return res
